Phase 0 (Hickey teardown): delete dead spine imitations + rename core_serve
Pure deletions of dead scaffolding that imitated the real ingest->reduce->
query->snapshot spine, plus one rename and one intra-crate method move. No
behavior change; verified by cargo build/clippy/test, swift build/test (1008
XCTest + 19 Swift Testing, 0 failures), the UniFFI drift guard, the structural
verifier, and a 3-lens adversarial review (deletion-safety: clean).

A1  Delete dead Rust projection/ + observation_journal scaffold
    - rm projection/mod.rs (project() = clone()+observations.len()); drop pub mod
    - rm storage/observation_journal.rs + its re-exports
    - rm the 3 dead observation types (ObservationSourceKind/Payload/Record) from
      observation/mod.rs -- PRESERVING the live transcript submodule
    - prune replay_diff.rs scaffold tests; keep determinism + deserialization tests
A2  Delete dead runtime/relocate.rs (#[allow(dead_code)] + unsafe libc volfs); the
    .project_path sidecar already supersedes it
A3  Delete mutate_idea/mutate_worktree STUB FFI methods + their 4 command types
    (faked ok:true while only bumping version + persisting a no-op snapshot);
    regenerated UniFFI bindings (capacitor_core.swift -328, both FFI headers)
C4  Delete dead SessionResolutionPolicy.sessionNameBelongsToProject heuristic +
    genericSessionNames + their tests; keep chooseSessionName (defers to Rust route)
D1  Delete the dead FFI-to-string encode island in RuntimeClient.swift (~300 LOC):
    SnapshotPayload(_ AppSnapshot) + 14 encode converters + 2 extension inits +
    8 enum->string + 2 Mirror helpers. Live HTTP-JSON decode path untouched;
    CoreRoutingTarget/RuntimeCaptureStatus type defs preserved
step6  Rename core_serve.rs -> core_diagnostics.rs (holds diagnostics, not serve
    dispatch); move the 7 idea-CRUD pass-throughs into new core_ideas.rs (FFI
    surface identical -- method names unchanged, no regen)

Also relabels the audit doc role audit->historical-evidence to satisfy the
doc-governance verifier (clears the one --evolve violation this branch added).

Pre-existing, NOT introduced here (separate housekeeping):
- observation::transcript scan_resolves_* test fails only under sandbox
  TMPDIR=/tmp/claude-501/.tmp* (transcript module is byte-identical to HEAD)
- cargo clippy --tests -D warnings debt in untouched method_runner test files
