+++ apps/swift/Sources/Capacitor/Bridge/capacitor_coreFFI.h
@@ -392,11 +392,6 @@ RustBuffer uniffi_capacitor_core_fn_method_coreruntime_load_ideas_order(void*_No
 RustBuffer uniffi_capacitor_core_fn_method_coreruntime_mutate_delegation(void*_Nonnull ptr, RustBuffer command, RustCallStatus *_Nonnull out_status
 );
 #endif
-#ifndef UNIFFI_FFIDEF_UNIFFI_CAPACITOR_CORE_FN_METHOD_CORERUNTIME_MUTATE_IDEA
-#define UNIFFI_FFIDEF_UNIFFI_CAPACITOR_CORE_FN_METHOD_CORERUNTIME_MUTATE_IDEA
-RustBuffer uniffi_capacitor_core_fn_method_coreruntime_mutate_idea(void*_Nonnull ptr, RustBuffer command, RustCallStatus *_Nonnull out_status
-);
-#endif
 #ifndef UNIFFI_FFIDEF_UNIFFI_CAPACITOR_CORE_FN_METHOD_CORERUNTIME_MUTATE_PROJECT
 #define UNIFFI_FFIDEF_UNIFFI_CAPACITOR_CORE_FN_METHOD_CORERUNTIME_MUTATE_PROJECT
 RustBuffer uniffi_capacitor_core_fn_method_coreruntime_mutate_project(void*_Nonnull ptr, RustBuffer command, RustCallStatus *_Nonnull out_status
@@ -405,11 +400,6 @@ RustBuffer uniffi_capacitor_core_fn_method_coreruntime_mutate_project(void*_Nonn
 #ifndef UNIFFI_FFIDEF_UNIFFI_CAPACITOR_CORE_FN_METHOD_CORERUNTIME_MUTATE_RUN
 #define UNIFFI_FFIDEF_UNIFFI_CAPACITOR_CORE_FN_METHOD_CORERUNTIME_MUTATE_RUN
 RustBuffer uniffi_capacitor_core_fn_method_coreruntime_mutate_run(void*_Nonnull ptr, RustBuffer command, RustCallStatus *_Nonnull out_status
-);
-#endif
-#ifndef UNIFFI_FFIDEF_UNIFFI_CAPACITOR_CORE_FN_METHOD_CORERUNTIME_MUTATE_WORKTREE
-#define UNIFFI_FFIDEF_UNIFFI_CAPACITOR_CORE_FN_METHOD_CORERUNTIME_MUTATE_WORKTREE
-RustBuffer uniffi_capacitor_core_fn_method_coreruntime_mutate_worktree(void*_Nonnull ptr, RustBuffer command, RustCallStatus *_Nonnull out_status
 );
 #endif
 #ifndef UNIFFI_FFIDEF_UNIFFI_CAPACITOR_CORE_FN_METHOD_CORERUNTIME_REMOVE_HOOKS
@@ -896,12 +886,6 @@ uint16_t uniffi_capacitor_core_checksum_method_coreruntime_mutate_delegation(voi
     
 );
 #endif
-#ifndef UNIFFI_FFIDEF_UNIFFI_CAPACITOR_CORE_CHECKSUM_METHOD_CORERUNTIME_MUTATE_IDEA
-#define UNIFFI_FFIDEF_UNIFFI_CAPACITOR_CORE_CHECKSUM_METHOD_CORERUNTIME_MUTATE_IDEA
-uint16_t uniffi_capacitor_core_checksum_method_coreruntime_mutate_idea(void
-    
-);
-#endif
 #ifndef UNIFFI_FFIDEF_UNIFFI_CAPACITOR_CORE_CHECKSUM_METHOD_CORERUNTIME_MUTATE_PROJECT
 #define UNIFFI_FFIDEF_UNIFFI_CAPACITOR_CORE_CHECKSUM_METHOD_CORERUNTIME_MUTATE_PROJECT
 uint16_t uniffi_capacitor_core_checksum_method_coreruntime_mutate_project(void
@@ -911,12 +895,6 @@ uint16_t uniffi_capacitor_core_checksum_method_coreruntime_mutate_project(void
 #ifndef UNIFFI_FFIDEF_UNIFFI_CAPACITOR_CORE_CHECKSUM_METHOD_CORERUNTIME_MUTATE_RUN
 #define UNIFFI_FFIDEF_UNIFFI_CAPACITOR_CORE_CHECKSUM_METHOD_CORERUNTIME_MUTATE_RUN
 uint16_t uniffi_capacitor_core_checksum_method_coreruntime_mutate_run(void
-    
-);
-#endif
-#ifndef UNIFFI_FFIDEF_UNIFFI_CAPACITOR_CORE_CHECKSUM_METHOD_CORERUNTIME_MUTATE_WORKTREE
-#define UNIFFI_FFIDEF_UNIFFI_CAPACITOR_CORE_CHECKSUM_METHOD_CORERUNTIME_MUTATE_WORKTREE
-uint16_t uniffi_capacitor_core_checksum_method_coreruntime_mutate_worktree(void
     
 );
 #endif
