+++ apps/swift/Tests/CapacitorTests/SessionResolutionPolicyTests.swift
@@ -143,78 +143,4 @@ final class SessionResolutionPolicyTests: XCTestCase {
         XCTAssertEqual(resolved, "Main",
                        "Generic session name matching should be case-insensitive")
     }
-
-    // MARK: - Direct Validation Method
-
-    func testSessionNameBelongsToProject() {
-        let policy = SessionResolutionPolicy()
-
-        // Same name
-        XCTAssertTrue(policy.sessionNameBelongsToProject(
-            sessionName: "circuit", projectPath: "/Code/circuit",
-        ))
-
-        // Generic
-        XCTAssertTrue(policy.sessionNameBelongsToProject(
-            sessionName: "main", projectPath: "/Code/circuit",
-        ))
-
-        // Session contains slug
-        XCTAssertTrue(policy.sessionNameBelongsToProject(
-            sessionName: "circuit-dev", projectPath: "/Code/circuit",
-        ))
-
-        // Slug contains session
-        XCTAssertTrue(policy.sessionNameBelongsToProject(
-            sessionName: "cap", projectPath: "/Code/capacitor",
-        ))
-
-        // Unrelated
-        XCTAssertFalse(policy.sessionNameBelongsToProject(
-            sessionName: "arc-design-studio", projectPath: "/Code/circuit",
-        ))
-
-        // Case insensitive
-        XCTAssertTrue(policy.sessionNameBelongsToProject(
-            sessionName: "Circuit", projectPath: "/Code/circuit",
-        ))
-    }
-
-    // MARK: - Short Slug Reverse Containment
-
-    func testRejectsSingleCharSessionNameForReverseContainment() {
-        let policy = SessionResolutionPolicy()
-        // "a" is contained in "capacitor" but should be rejected (too short for reverse match)
-        XCTAssertFalse(policy.sessionNameBelongsToProject(
-            sessionName: "a",
-            projectPath: "/Code/capacitor",
-        ))
-    }
-
-    func testRejectsTwoCharSessionNameForReverseContainment() {
-        let policy = SessionResolutionPolicy()
-        // "go" is contained in "golang" but should be rejected (too short for reverse match)
-        XCTAssertFalse(policy.sessionNameBelongsToProject(
-            sessionName: "go",
-            projectPath: "/Code/golang",
-        ))
-    }
-
-    func testAcceptsTwoCharSessionNameWhenExactMatch() {
-        let policy = SessionResolutionPolicy()
-        // "go" contains "go" via forward containment (exact match), so this should pass
-        XCTAssertTrue(policy.sessionNameBelongsToProject(
-            sessionName: "go",
-            projectPath: "/Code/go",
-        ))
-    }
-
-    func testAcceptsThreeCharSessionNameForReverseContainment() {
-        let policy = SessionResolutionPolicy()
-        // "cap" is >= 3 chars and is contained in "capacitor", so reverse match succeeds
-        XCTAssertTrue(policy.sessionNameBelongsToProject(
-            sessionName: "cap",
-            projectPath: "/Code/capacitor",
-        ))
-    }
 }
