+++ core/capacitor-core/src/core_diagnostics.rs
