+++ core/capacitor-core/src/core_ingest.rs
@@ -228,50 +228,6 @@ impl CoreRuntime {
         Ok(outcome)
     }
 
-    /// Accept an idea mutation command.
-    ///
-    /// STUB: logs the command and persists a version bump, but does not yet
-    /// mutate model state. Callers receive `ok: true` as acknowledgment.
-    pub fn mutate_idea(
-        &self,
-        command: MutateIdeaCommand,
-    ) -> Result<MutationOutcome, CoreRuntimeError> {
-        let mut state = self.lock_state()?;
-        state.events_ingested = state.events_ingested.saturating_add(1);
-        let message = format!(
-            "idea mutation accepted kind={:?} project_path={} idea_id={} (stub)",
-            command.kind, command.project_path, command.idea_id
-        );
-        let outcome = MutationOutcome { ok: true, message };
-        self.bump_version_and_notify();
-        let snapshot = state.snapshot();
-        drop(state);
-        self.persist_snapshot(&snapshot)?;
-        Ok(outcome)
-    }
-
-    /// Accept a worktree mutation command.
-    ///
-    /// STUB: logs the command and persists a version bump, but does not yet
-    /// mutate model state. Callers receive `ok: true` as acknowledgment.
-    pub fn mutate_worktree(
-        &self,
-        command: MutateWorktreeCommand,
-    ) -> Result<MutationOutcome, CoreRuntimeError> {
-        let mut state = self.lock_state()?;
-        state.events_ingested = state.events_ingested.saturating_add(1);
-        let message = format!(
-            "worktree mutation accepted kind={:?} repo_path={} worktree_name={} force={} (stub)",
-            command.kind, command.repo_path, command.worktree_name, command.force
-        );
-        let outcome = MutationOutcome { ok: true, message };
-        self.bump_version_and_notify();
-        let snapshot = state.snapshot();
-        drop(state);
-        self.persist_snapshot(&snapshot)?;
-        Ok(outcome)
-    }
-
     pub fn mutate_delegation(
         &self,
         command: MutateDelegationCommand,
@@ -296,98 +252,5 @@ impl CoreRuntime {
         drop(state);
         self.persist_snapshot(&snapshot)?;
         Ok(outcome)
-    }
-
-    pub fn capture_idea(
-        &self,
-        project_path: String,
-        idea_text: String,
-    ) -> Result<String, CoreRuntimeError> {
-        runtime::ideas::capture_idea_with_storage(&self.app_storage, &project_path, &idea_text)
-            .map_err(|error| CoreRuntimeError::from(error.to_string()))
-    }
-
-    pub fn update_idea_status(
-        &self,
-        project_path: String,
-        idea_id: String,
-        new_status: String,
-    ) -> Result<(), CoreRuntimeError> {
-        runtime::ideas::update_idea_status_with_storage(
-            &self.app_storage,
-            &project_path,
-            &idea_id,
-            &new_status,
-        )
-        .map_err(|error| CoreRuntimeError::from(error.to_string()))
-    }
-
-    pub fn update_idea_effort(
-        &self,
-        project_path: String,
-        idea_id: String,
-        new_effort: String,
-    ) -> Result<(), CoreRuntimeError> {
-        runtime::ideas::update_idea_effort_with_storage(
-            &self.app_storage,
-            &project_path,
-            &idea_id,
-            &new_effort,
-        )
-        .map_err(|error| CoreRuntimeError::from(error.to_string()))
-    }
-
-    pub fn update_idea_triage(
-        &self,
-        project_path: String,
-        idea_id: String,
-        new_triage: String,
-    ) -> Result<(), CoreRuntimeError> {
-        runtime::ideas::update_idea_triage_with_storage(
-            &self.app_storage,
-            &project_path,
-            &idea_id,
-            &new_triage,
-        )
-        .map_err(|error| CoreRuntimeError::from(error.to_string()))
-    }
-
-    pub fn update_idea_title(
-        &self,
-        project_path: String,
-        idea_id: String,
-        new_title: String,
-    ) -> Result<(), CoreRuntimeError> {
-        runtime::ideas::update_idea_title_with_storage(
-            &self.app_storage,
-            &project_path,
-            &idea_id,
-            &new_title,
-        )
-        .map_err(|error| CoreRuntimeError::from(error.to_string()))
-    }
-
-    pub fn update_idea_description(
-        &self,
-        project_path: String,
-        idea_id: String,
-        new_description: String,
-    ) -> Result<(), CoreRuntimeError> {
-        runtime::ideas::update_idea_description_with_storage(
-            &self.app_storage,
-            &project_path,
-            &idea_id,
-            &new_description,
-        )
-        .map_err(|error| CoreRuntimeError::from(error.to_string()))
-    }
-
-    pub fn save_ideas_order(
-        &self,
-        project_path: String,
-        idea_ids: Vec<String>,
-    ) -> Result<(), CoreRuntimeError> {
-        runtime::ideas::save_ideas_order_with_storage(&self.app_storage, &project_path, idea_ids)
-            .map_err(|error| CoreRuntimeError::from(error.to_string()))
     }
 }
